Fixed bug with preprocessor not correctly removing multilines
This issue was two bugs, really. We can now remove multilines in
multilines. The second bug was related to removing a multiline with a
`/` inside, which caused the comment-removal loop to escape before it
found the end of the multiline comment (this meant we had only
partially removed the multiline in these cases!!!)